Dockerized-Django/app/requirements.txt
Django
psycopg2-binary==2.9.1
djangorestframework
django-cors-headers
